Fix build errors and warnings on Windows.

+++ src/detail/ckms_quantiles.cc
@@ -68,7 +68,7 @@ void CKMSQuantiles::reset() {
 
 double CKMSQuantiles::allowableError(int rank) {
   auto size = sample_.size();
-  double minError = size + 1;
+  double minError = double(size + 1);
 
   for (const auto& q : quantiles_.get()) {
     double error;
@@ -116,7 +116,7 @@ bool CKMSQuantiles::insertBatch() {
     if (idx - 1 == 0 || idx + 1 == sample_.size()) {
       delta = 0;
     } else {
-      delta = static_cast<int>(std::floor(allowableError(idx + 1))) + 1;
+      delta = static_cast<int>(std::floor(allowableError(int(idx + 1)))) + 1;
     }
 
     sample_.emplace(sample_.begin() + idx, v, 1, delta);
@@ -142,7 +142,7 @@ void CKMSQuantiles::compress() {
     next = idx++;
 
     if (sample_[prev].g + sample_[next].g + sample_[next].delta <=
-        allowableError(idx - 1)) {
+        allowableError(int(idx - 1))) {
       sample_[next].g += sample_[prev].g;
       sample_.erase(sample_.begin() + prev);
     }
